requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.1
lightgbm>=3.3
pyyaml>=6.0
geopandas>=0.12
rasterio>=1.3
xarray>=2023.10.0
pyproj>=3.5
shapely>=1.8
joblib>=1.2
pytest>=7.0
streamlit>=1.28
matplotlib>=3.7
shap>=0.42
